+++ src/cli/commands/public-cmds/bind-cmd.js
@@ -7,11 +7,16 @@ export default class Create extends Command {
   name = 'bind';
   description = 'Call the driver bind action';
   alias = 'b';
-  opts = [];
+  opts = [
+    ['v', 'verbose', 'showing the driver path'],
+  ];
   loader = true;
 
-  action(): Promise<*> {
+  action(args: string[], { verbose }: { verbose: ?bool }): Promise<*> {
     return getDriver().then((driverObj) => {
+      // TODO: this is a quick hack to get the driver path for debugging purposes
+      // it should be part of the logging once we implement "verbose" for all commands
+      if (verbose) console.log('Driver path', driverObj.driverPath()); // eslint-disable-line
       return driverObj.getDriver(false).bind({});
     });
   }